Another attempt to get the "attempted completion" stuff to work, and re-adding CompletionMatches function (that works - looks like I really need to do all the function pointer stuff in C)
Now you can set the attempted completion function and, for example, call CompletionMatches for the first work (command name) and maybe with a different callback for the rest of the line.

Still can't get the documented readline behaviour (if the attempted completion function return NULL, the default function is called). Once I set it, I lose the default :(